Pending changes exported from your codespace

+++ CH02/02_12/02_12-typecast1.c
@@ -2,12 +2,12 @@
 
 int main()
 {
-	int a,b;
+	int a, b;
 
 	a = 100;
 	b = 7;
 
-	printf("%d / %d = %d\n",a,b,a/b);
+	printf("%d / %d = %.3f\n", a, b, (float)a / b);
 
-	return(0);
+	return (0);
 }
